@@ -6,7 +6,7 @@
 }
 
 
-
+    <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.css" />
 <div class="home-page-background">
     <div class="content">
         @using (Html.BeginForm("Search", "Home", FormMethod.Post))
@@ -31,16 +31,16 @@
     <div class="col-md-4">
         <h5>LEARN MORE</h5>
         <div style="font-size:16px;">
-            <a href="https://azure.microsoft.com/en-us/services/search/">Azure Search</a>
+            <a href="https://docs.microsoft.com/en-us/azure/search/search-what-is-azure-search">Azure Cognitive Search</a>
         </div>
         <div style="font-size:16px;">
-            <a href="https://docs.microsoft.com/en-us/azure/search/cognitive-search-resources-documentation">Cognitive Search</a>
+            <a href="https://docs.microsoft.com/en-us/azure/cognitive-services/Welcome">Cognitive Services</a>
         </div>
         <div style="font-size:16px;">
-            <a href="https://azure.microsoft.com/en-us/services/cognitive-services/text-analytics/">Text Analytics</a>
+            <a href="https://docs.microsoft.com/en-us/azure/devops/pipelines/targets/azure-machine-learning?view=azure-devops&tabs=yaml">Azure Machine Learning</a>
         </div>
         <div style="font-size:16px;">
-            <a href="https://azure.microsoft.com/en-us/services/cognitive-services/computer-vision/">Computer Vision</a>
+            <a href="https://azure.microsoft.com/en-us/resources/knowledge-mining-the-next-wave-of-artificial-intelligence-led-transformation/">Knowledge Mining</a>
         </div>
     </div>
 </div>
